Improved upstart script
After a server reboot, mesos could start before network is up.
If the server has local ips such as 127.0.1.1, mesos processes will
adversise an ipaddress 127.0.1.1 to slaves (or master) which lead to
chaos in the cluster.

http://serverfault.com/a/370869/98600 suggests to use a better start
condition to ensure that mesos processes are started on correct runlevel
(after network is up).
Related documentation for this is on
http://upstart.ubuntu.com/cookbook/#really-understanding-start-on-and-stop-on